requirements.txt
django

Pillow

django-simple-history

mysqlclient